Expose the ability to store and retrieve a datatype from the TensorShape representation to the Tensor class, and use this ability to store the DataType in the extra byte that is unused inside the TensorShape representation.  This makes sizeof(Tensor) == 32 bytes on a machine with 8-byte pointers, instead of 40 bytes.
Change the value field in AllocatorAttributes from uint32 to uint8, and then
rearranged the order of the fields in ExecutorState::Entry so that this fits
in the same word as the has_value boolean.  After the Tensor change above,
Entry was 72 bytes, and this change makes it 64 bytes.
Change: 114358788

+++ tensorflow/core/framework/allocator.h
@@ -240,7 +240,7 @@ struct AllocatorAttributes {
 
   void Merge(AllocatorAttributes other) { value |= other.value; }
 
-  uint32 value = 0;
+  uint8 value = 0;
 };
 
 // Returns a trivial implementation of Allocator which uses the system
